requirements.txt
loguru~=0.6.0
boto3~=1.26.72
PyYAML~=5.4.1
